requirements-deploy.txt
torch==2.5.1
torchvision==0.20.1
open_clip_torch
transformers
accelerate
Pillow
requests
tqdm
streamlit
plotly
wandb
jupyter
umap-learn
matplotlib
seaborn
scikit-learn
huggingface_hub
